backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
librosa==0.10.1
soundfile==0.12.1
tensorflow==2.15.0
facenet-pytorch==2.5.3
transformers==4.37.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
aiosqlite==0.19.0
